Move folder rename and change location API to a separate module
