+++ src/http/photos_api/mod.rs
@@ -1,7 +1,6 @@
 mod favorite;
 mod sync;
-
-use std::string::ToString;
+mod move_photos;
 
 use axum::{
     Json, Router,
@@ -26,16 +25,14 @@ use time::serde::timestamp;
 
 pub fn router(app_state: AppStateRef) -> Router {
     Router::new()
-        .route("/", get(photos_list))
         .nest("/sync", sync::router())
+        .nest("/move", move_photos::router())
         .route("/duplicates", get(get_duplicates))
         .route("/download/{photo_id}", get(download_photo))
         .route("/preview/{photo_id}", get(preview_photo))
         .route("/exif/{photo_id}", get(get_photo_exif))
         .route("/upload", post(upload_photo))
         .route("/delete/{photo_id}", delete(delete_photo))
-        .route("/change_location/{photo_id}", post(change_photo_location))
-        .route("/rename_folder", post(rename_folder))
         .nest("/favorite", favorite::router())
         .with_state(app_state)
 }
@@ -48,22 +45,6 @@ fn check_has_access(user: Option<User>, photo: &Photo) -> Result<User, StatusCod
     } else {
         Err(StatusCode::FORBIDDEN)
     }
-}
-
-async fn photos_list(
-    State(state): State<AppStateRef>,
-    auth: AuthSession,
-) -> AxumResult<impl IntoResponse> {
-    let user = auth.user.ok_or(StatusCode::BAD_REQUEST)?;
-
-    Ok(Json(
-        state
-            .photos_repo
-            .get_photos_by_user_and_public(user.id)
-            .await
-            .map_err(internal_error)?
-            .photos,
-    ))
 }
 
 async fn get_duplicates(
@@ -166,7 +147,6 @@ async fn get_photo_exif(
 }
 
 #[derive(Debug, serde::Deserialize)]
-#[serde(rename_all = "camelCase")]
 struct UploadDataQuery {
     #[serde(with = "timestamp")]
     time_created: OffsetDateTime,
@@ -269,133 +249,4 @@ async fn delete_photo(
         .map_err(internal_error)?;
 
     Ok(())
-}
-
-#[derive(serde::Deserialize)]
-struct ChangeLocationQuery {
-    make_public: bool,
-    target_folder_name: Option<String>,
-}
-
-async fn change_photo_location(
-    State(state): State<AppStateRef>,
-    Path(photo_id): Path<i64>,
-    Query(query): Query<ChangeLocationQuery>,
-    auth: AuthSession,
-) -> AxumResult<impl IntoResponse> {
-    let storage = &state.storage;
-    let photo = state
-        .photos_repo
-        .get_photo(photo_id)
-        .await
-        .map_err(internal_error)?;
-    let user = check_has_access(auth.user, &photo)?;
-
-    let target_user_name = if query.make_public {
-        PUBLIC_USER_ID.to_string()
-    } else {
-        user.id
-    };
-
-    let source_path = photo.partial_path();
-    let changed_photo = Photo {
-        id: photo.id(),
-        user_id: target_user_name,
-        name: photo.name,
-        created_at: photo.created_at,
-        file_size: photo.file_size,
-        folder: query.target_folder_name,
-    };
-    let destination_path = changed_photo.partial_path();
-
-    if source_path == destination_path {
-        return Err((
-            StatusCode::BAD_REQUEST,
-            "Source and destination are the same",
-        )
-            .into_response()
-            .into());
-    }
-
-    info!("Moving photo from {source_path} to {destination_path}");
-
-    storage
-        .move_photo(&source_path, &destination_path)
-        .map_err(|e| {
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                format!("Failed moving the photo: {e}"),
-            )
-        })?;
-
-    state
-        .photos_repo
-        .update_photo(&changed_photo)
-        .await
-        .inspect_err(|e| {
-            error!("Failed to update photo: {e}");
-            // Try to undo the move
-            let _ = storage.move_photo(&destination_path, &source_path);
-        })
-        .map_err(internal_error)?;
-
-    Ok(Json(changed_photo))
-}
-
-#[derive(serde::Deserialize)]
-struct RenameFolderQuery {
-    source_is_public: bool,
-    source_folder_name: String,
-    target_make_public: bool,
-    target_folder_name: Option<String>,
-}
-
-async fn rename_folder(
-    State(state): State<AppStateRef>,
-    Query(query): Query<RenameFolderQuery>,
-    auth: AuthSession,
-) -> AxumResult<impl IntoResponse> {
-    let user = auth.user.expect("User should be logged in");
-    let storage = &state.storage;
-
-    let source_user_name = if query.source_is_public {
-        PUBLIC_USER_ID
-    } else {
-        &user.id
-    };
-    let target_user_name = if query.target_make_public {
-        PUBLIC_USER_ID
-    } else {
-        &user.id
-    };
-
-    let photos_to_move = state
-        .photos_repo
-        .get_photos_in_folder(source_user_name, query.source_folder_name)
-        .await
-        .map_err(internal_error)?;
-    let mut moved_photos = Vec::with_capacity(photos_to_move.len());
-
-    for mut photo in photos_to_move {
-        let source_path = photo.partial_path();
-
-        photo.user_id = target_user_name.to_owned();
-        photo.folder = query.target_folder_name.clone();
-        let destination_path = photo.partial_path();
-
-        if let Err(e) = storage.move_photo(&source_path, &destination_path) {
-            warn!("Failed to move the photo: {e}");
-            continue;
-        }
-
-        if let Err(e) = state.photos_repo.update_photo(&photo).await {
-            // If the database operation failed for some reason, try to move the image back
-            error!("Failed to update the photo: {e:?}");
-            let _ = storage.move_photo(&destination_path, &source_path);
-        }
-
-        moved_photos.push(photo);
-    }
-
-    Ok(Json(moved_photos))
 }
